refactor(core): split browser-safe SDK and tooling entrypoints
Keep the default zodiac-core entrypoint limited to browser-safe SDK helpers and move artifact/deployment utilities behind the tooling entrypoint. Update artifact type imports and tests to use the new tooling path.
