+++ package.json
@@ -8,11 +8,23 @@
   "main": "./dist/cjs/index.js",
   "module": "./dist/esm/index.js",
   "types": "./dist/esm/index.d.ts",
+  "typesVersions": {
+    "*": {
+      "tooling": [
+        "dist/esm/tooling.d.ts"
+      ]
+    }
+  },
   "exports": {
     ".": {
       "types": "./dist/esm/index.d.ts",
       "import": "./dist/esm/index.js",
       "require": "./dist/cjs/index.js"
+    },
+    "./tooling": {
+      "types": "./dist/esm/tooling.d.ts",
+      "import": "./dist/esm/tooling.js",
+      "require": "./dist/cjs/tooling.js"
     },
     "./contracts/*": "./contracts/*",
     "./package.json": "./package.json"
