+++ src/artifact/verifyAllMastercopies.ts
@@ -3,7 +3,7 @@ import { existsSync, readFileSync } from "fs";
 import { defaultMastercopyArtifactsFile } from "./internal/paths.js";
 import { verifySourceCode } from "./internal/etherscan.js";
 
-import { MastercopyArtifact } from "../types.js";
+import { MastercopyArtifact } from "./types.js";
 
 /**
  * Iterates through each entry in the mastercopy artifacts file and verifies the mastercopy on an Etherscan-compatible block explorer.
