+++ src/tooling.ts
@@ -1,0 +1,23 @@
+export * from "./index.js";
+export type { BuildArtifact, MastercopyArtifact } from "./artifact/types.js";
+
+// tooling
+import deployFactories from "./tooling/deployFactories.js";
+import deployMastercopy from "./tooling/deployMastercopy.js";
+import deployProxy from "./tooling/deployProxy.js";
+import verifyMastercopy from "./tooling/verifyMastercopy.js";
+
+// artifact
+import readMastercopies from "./artifact/readMastercopies.js";
+import writeMastercopyFromBuild from "./artifact/writeMastercopyFromBuild.js";
+import writeMastercopyFromExplorer from "./artifact/writeMastercopyFromExplorer.js";
+
+export {
+  deployFactories,
+  deployMastercopy,
+  deployProxy,
+  readMastercopies,
+  verifyMastercopy,
+  writeMastercopyFromBuild,
+  writeMastercopyFromExplorer,
+};
